Splatoon 2 and 3 fully functional on the watch

+++ SplatInfo/Splatoon2InkAPI.swift
@@ -235,7 +235,7 @@ class Splatoon2InkAPI {
     }
 
     struct ModeAPIResponse: Codable {
-        let id : Int
+        let id : Int64
         let stageA: StageAPIResponse
         let stageB: StageAPIResponse
         let startTime: Date
